fix errors

+++ budget-optimizer-backend/src/main/java/com/budgetoptimizer/budget_optimizer_backend/service/UsuarioService.java
@@ -390,7 +390,7 @@ public class UsuarioService {
 
                 // Beneficios
                 .descuentoPorcentaje(
-                        accountType.getDescuentoPorcentaje()
+                        BigDecimal.valueOf(accountType.getDescuentoPorcentaje())
                 )
 
                 .tieneBeneficios(
